refactor ui module
